feat(sdk-spec): Krok 4a — strict canonical CBOR validator (Faza 6 deterministic profile, RFC 8949 §4.2.1 + f64-only) + MAX_NESTING_DEPTH=64 anti-DoS + usize::try_from for wasm32 + f32→f64 migration (GraphEdge.weight, FileMeta.upload_progress, AspectRatio::Custom.ratio, Viewport.density)

+++ tentaflow-sdk-spec/src/protocol/ui/panel.rs
@@ -20,7 +20,7 @@ pub struct Viewport {
     #[n(1)]
     pub height_px: u32,
     #[n(2)]
-    pub density: f32,
+    pub density: f64,
 }
 
 /// Panel-open context enriched by core (assigned_epoch is core-controlled).
